% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compare_element}
\alias{compare_element}
\title{Compare an element to elements in a vector}
\usage{
compare_element(element, vector)
}
\arguments{
\item{element}{Something to compare to. Can be any single element RObject}

\item{vector}{A vector of things to compare element to. Can be any types.}
}
\value{
An (arma) indicator vector. 1 means element same as vector's element and 0 otherwise.
}
\description{
Compare an element to elements in a vector, independent of data type.
}
\details{
Sometime when comparing RObjects of arbitrary type doesn't work right because the wrong type
is guessed, i.e. element RObject cast to a hex code for a char but needs it to be interpreted as a hex
code for a string. This function just casts EVERYTHING to char hex codes using Rf_asChar(). This trick
seems to work.
}
\examples{
XXXX

}
