% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{compute.model.matrix.e}
\alias{compute.model.matrix.e}
\title{Compute model matrix for a set of configurations}
\usage{
compute.model.matrix.e(configs.mat, crf)
}
\arguments{
\item{configs.mat}{a set of configurations}

\item{crf}{crf object}
}
\value{
a MRF model matrix
}
\description{
Compute model matrix for a set of configurations
}
\details{
This version assumes configurations are given in terms of node state indices.
}
\examples{
XXXX

}
