% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{config.energy.e}
\alias{config.energy.e}
\title{Efficient energy function for a configuration of states}
\usage{
config.energy.e(config, edges.mat, one.nlp, two.nlp)
}
\arguments{
\item{config}{a node configuration (configuration state) vector}

\item{edges.mat}{matrix of connected node edges}

\item{two.lgp}{negative-log edge potentials (two-body "energies"), entered as a list}
}
\value{
The function will XX
}
\description{
Compute total energy of a configuration. Uses efficient one and two body energy functions.
Assumes node/edge energies are both entered as lists.
}
\details{
Uses efficient one and two body energy functions. No need to pass in feature function.
Assumes node/edge energies are both entered as lists.
}
