% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{config.energy}
\alias{config.energy}
\title{Energy function for a configuration of states}
\usage{
config.energy(config, edges.mat, one.nlp, two.nlp, ff)
}
\arguments{
\item{config}{a node configuration (configuration state) vector}

\item{edges.mat}{matrix of connected node edges}

\item{ff}{The feature function}

\item{two.lgp}{negative-log edge potentials (two-body "energies"), entered as a list}

\item{...}{Optional arguments for feature function.}
}
\value{
The function will XX
}
\description{
Compute total energy of a configuration. Assumes node/edge energies are both entered as lists.
}
\details{
Assumes node/edge energies are both entered as lists.
}
