% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{config_energy_e_C}
\alias{config_energy_e_C}
\title{Efficient configuration energy function}
\usage{
config_energy_e_C(config, edges_mat, one_nlp, two_nlp)
}
\arguments{
\item{yA}{node A state index.}

\item{yB}{node B state index.}

\item{wAB}{omega matrix}
}
\description{
Efficient configuration energy function. Uses above efficient one and two
body energy functions. No feature function required.
}
