% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{configs.n2i}
\alias{configs.n2i}
\title{Convert name(s) to index(ies) for a matrix of configurations}
\usage{
configs.n2i(configs.mat, crf, node.col.lblQ = T)
}
\arguments{
\item{configs.mat}{A matrix of configurations}

\item{crf}{A crf object}

\item{node.col.lblQ}{Whether or not to label columns with node names}
}
\value{
Configuration matrix in terms of node state indices
}
\description{
Convert name(s) to index(ies) for a matrix of configurations
}
\details{
Function converts node state names in configurations to node state indices.
The biggest reason to do this is that the more efficient energy functions can only use
node state indices.
}
