% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{distribution.from.energies}
\alias{distribution.from.energies}
\title{Compute the joint distribution from the node and edge energies}
\usage{
distribution.from.energies(
  crf = NULL,
  node.energies = NULL,
  edge.energies = NULL,
  use.gRQ = F
)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
\details{
The function will XXXX
}
