% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_util.R
\name{distribution.from.potentials}
\alias{distribution.from.potentials}
\title{Compute the joint distribution from the node and edge potentials using gRbase functions}
\usage{
distribution.from.potentials(gRbase.node.potentials, gRbase.edge.potentials)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will state probalities in gRbase table form as well as logZ
}
\description{
The function will XXXX
}
\details{
The function will XXXX
}
