% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_util.R
\name{dress.sample}
\alias{dress.sample}
\title{Dress a sample with node and state names in crf object}
\usage{
dress.sample(crf, samples)
}
\arguments{
\item{crf}{A crf object which also contains node and state names (i.e. out augmented crf object)}

\item{samples}{The samples generated by one of the CRF sampling functions}
}
\value{
The sample dressed with node and state names contained in the crf object.
}
\description{
Dress a sample with node and state names in crf object
}
\details{
The function will dress up a sample configuration matrix with node names as the columns
and state names for the configuration outcome elements. The sampling functions from CRF
(e.g. sample.junction()) just spit out an undressed matrix of samples with sample state numbers
(e.g. 1, 2, ....) for the configuration elements. This is needed sometimes and often more
convenient, but for diplay purposes the user may want more labels on the sample space. Hence this
(cosmetic) function.
}
\examples{
XXXX

}
