% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{dump.crf}
\alias{dump.crf}
\title{Print out all contents of a crf object}
\usage{
dump.crf(crf)
}
\arguments{
\item{crf}{A crf object}
}
\value{
Nothing.
}
\description{
Print out all contents of a crf object
}
\details{
The function dumps out the contents of a crf object
}
\examples{
XX

}
