% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{energyeC}
\alias{energyeC}
\title{C Energy function wrapper}
\usage{
energyeC(config, crf)
}
\arguments{
\item{config}{a configuration}

\item{crf}{a crf object}
}
\value{
Energy of the configuration.
}
\description{
Wrapper for C version of the energy function
}
\details{
The function is a wrapper for config_energy_e_C. Easier input. **NOTE: assumes
state names contained in the crf object are positive integers, i.e. 1, 2, 3, ..., etc.
}
\examples{
XXXX

}
