% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{energyf}
\alias{energyf}
\title{Energy function wrapper}
\usage{
energyf(config, crf, ff, ...)
}
\arguments{
\item{config}{a configuration}

\item{crf}{a crf object}

\item{ff}{the feature function}

\item{...}{optional arguments for feature function.}
}
\value{
Energy of the configuration.
}
\description{
Wrapper for energy function
}
\details{
The function is a wrapper for config.energy. Easier input.
}
\examples{
XXXX

}
