% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{energyphe}
\alias{energyphe}
\title{Energy function wrapper}
\usage{
energyphe(config, crf)
}
\arguments{
\item{config}{a configuration}

\item{crf}{a crf object}
}
\value{
Energy of the configuration.
}
\description{
Wrapper for energy function which uses phi vector for a config to compute energy
}
\details{
Compute energy using phi vector for configuration: E({\bf X}) = {\boldsymbol \theta}^{\dagger} {\boldsymbol \phi}({\bf X}) **NOTE: assumes
state names contained in the crf object are positive integers, i.e. 1, 2, 3, ..., etc.
}
\examples{
XXXX

}
