% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{Eone.e}
\alias{Eone.e}
\title{Efficient one-body (node) energy function}
\usage{
Eone.e(yA, tA)
}
\arguments{
\item{yA}{a node state}

\item{tA}{tau vector}
}
\value{
One-body (node) energy.
}
\description{
Efficient one-body (node) energy function
}
\details{
Efficient one-body (node) energy function. No feature function (ff) to pass in. Just does
the energy calculation as a look-up and not a matrix multiply. Assumes state names are just INDICES
(i.e. yA = 1, 2, 3, ......) and all state weights are 1.
}
\examples{
XXXX

}
