% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Eone_C}
\alias{Eone_C}
\title{One-body (node) energy function}
\usage{
Eone_C(yA, tA, ff, dots = NULL)
}
\arguments{
\item{yA}{a node state}

\item{tA}{tau vector}

\item{ff}{feature function}

\item{dots}{list of optional arguments for feature function}
}
\value{
One-body (node) energy.
}
\description{
One-body (node) energy function
}
\details{
One-body (node) energy function
}
\examples{
# node weights:
tau2 <- c( 1, -1.3)
tau3 <- c( 1, -1.3, 0.1)

Eone_C(yA = 2, tA = tau3, ff = ff1_C, list(nss.dim=3))
Eone_C(yA = 3, tA = tau3, ff = ff1_C, list(nss.dim=3))
Eone_C(yA = 2, tA = tau2, ff = ff1_C)
Eone_C(yA = "a", tA = tau3, ff = ff1_C, list(nss.vec = c("a", "b", "c")))
Eone_C(yA = 2, tA = tau2, ff = ff0_C)
Eone_C(yA = 1, tA = tau2, ff = ff0_C)
Eone_C(yA = 3, tA = tau2, ff = ff0_C) # Throws error because less general ff0 used

}
