% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{Etwo.e}
\alias{Etwo.e}
\title{Efficient two-body (edge) energy function}
\usage{
Etwo.e(yA, yB, wAB)
}
\arguments{
\item{yA}{node A state}

\item{yB}{node B state}

\item{wAB}{omega matrix}
}
\value{
Two-body (edge) energy.
}
\description{
Efficient two-body (edge) energy function
}
\details{
Efficient two-body (edge) energy function. No feature function (ff) to pass in. Just does
the energy calculation as a look-up and not a matrix multiply. Assumes state names are just INDICES
(i.e. yA , yB= 1, 2, 3, ......) and all state weights are 1.
}
\examples{
XXXX

}
