% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{Etwo}
\alias{Etwo}
\title{Two-body (edge) energy function}
\usage{
Etwo(yA, yB, wAB, ff)
}
\arguments{
\item{yA}{A node state}

\item{yB}{A node state}

\item{wAB}{Tau vector}

\item{ff}{Feature function}
}
\value{
Two-body (edge) energy.
}
\description{
Two-body (edge) energy function
}
\details{
Two-body (edge) energy function.
}
\examples{
XXXXX

}
