% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Etwo_C}
\alias{Etwo_C}
\title{Two-body (edge) energy function}
\usage{
Etwo_C(yA, yB, wAB, ff, dots = NULL)
}
\arguments{
\item{yA}{node A state}

\item{yB}{node B state}

\item{wAB}{omega matrix}

\item{ff}{feature function}

\item{dots}{list of optional arguments for feature function}
}
\value{
Two-body (edge) energy.
}
\description{
Two-body (edge) energy function
}
\details{
Two-body (edge) energy function
}
\examples{
# edge weights:
omega22 <- rbind(
 c( 3.5, -1.4),
 c(-1.4,  2.5)
)

omega33 <- rbind(
 c( 3.5, -1.4,   2.5),
 c(-1.4,  2.5,  -0.8),
 c( 3.8,  0.95, -7.6)
)

Etwo_C(yA = 2, yB = 1, wAB = omega22, ff = ff1_C, list(nss.dim=2))
Etwo_C(yA = 2, yB = 1, wAB = omega22, ff = ff1_C, list(nss.dim=3)) # Throws error
Etwo_C(yA = 2, yB = 2, wAB = omega33, ff = ff1_C, list(nss.dim=3)) # Throws error
Etwo_C(yA = "c", yB = "b", wAB = omega33, ff = ff1_C, list(nss.vec=c("a", "b", "c")))

Etwo_C(yA = 2, yB = 1, wAB = omega22, ff = ff0_C)
Etwo_C(yA = 2, yB = 1, wAB = omega33, ff = ff0_C)
Etwo_C(yA = 2, yB = 1, wAB = omega33, ff = ff1_C) # Throws error bec nss.dim=2 assumed
Etwo_C(yA = 2, yB = 1, wAB = omega22, ff = ff1_C)

}
