% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Etwo_e_C}
\alias{Etwo_e_C}
\title{Efficient two-body (edge) energy function}
\usage{
Etwo_e_C(yA, yB, wAB)
}
\arguments{
\item{yA}{node A state index.}

\item{yB}{node B state index.}

\item{wAB}{omega matrix}
}
\description{
Efficient two-body (edge) energy function. No feature function required.
}
