% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{exp_list}
\alias{exp_list}
\title{Exp a bunch of stuff (vectors, matrices, etc) in a list}
\usage{
exp_list(a.list.for.eing, negexpQ = F)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Handy for going from node/edge energies (-log potentials) in list to node/edge potentials
}
\details{
The function will XXXX
}
