% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff0.R
\name{ff0}
\alias{ff0}
\title{Feature function}
\usage{
ff0(st, ss.dim = NULL, w.vec = NULL, st.vec = NULL)
}
\arguments{
\item{st}{State label}

\item{ss.dim}{State space dimension. Default is 2 (Ising)}

\item{w.vec}{Optional weight vector}

\item{st.vec}{Optional vector for state names}
}
\value{
The feature function, which is a vector.
}
\description{
The function is a handy feature function for testing.
}
\details{
The function is a handy feature function for testing. It can handle both Ising and Potts state spaces.
}
\examples{
ff0(st = 3, ss.dim = 4, st.vec = c(0,1,2,3))
ff0(st = "a", st.vec = c("a","b"))
ff0(st = "ID", ss.dim = 3, st.vec = c("ID","exclude","inconclusive"))
ff0(1)
ff0(1,3)
ff0(st = "b")           # Throws error
ff0(st = 5)             # Throws error
ff0(st = 5, ss.dim = 4) # Throws error

}
