% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ff1.R
\name{ff1}
\alias{ff1}
\title{Feature function}
\usage{
ff1(ns, ...)
}
\arguments{
\item{ns}{state label of node (\emph{i.e.} node state)}

\item{...}{Optional arguments: \cr
\describe{
  \item{\code{nss.vec}}{node state space (a vector)}
  \item{\code{nss.dim}}{node state space dimension}
  \item{\code{w.vec}}{node state space weight vector}
}}
}
\value{
The feature function, which is a vector.
}
\description{
The function is a handy feature function for testing.
}
\details{
The function is a handy feature function for testing. It can handle both Ising and Potts state spaces.
 NOTE: All optional arguments must be labeled with their name.
}
\examples{
ff1(ns = 3, nss.dim = 4, nss.vec = c(0,1,2,3))
ff1(ns = "a", nss.vec = c("a","b"))
ff1(ns = "ID", nss.dim = 3, nss.vec = c("ID","exclude","inconclusive"))
ff1(1)
ff1(1,3)                 # Ambiguous. Optional arguments must be labeled with their name.
ff1(ns = "b")            # Throws error
ff1(ns = 5)              # Throws error
ff1(ns = 5, nss.dim = 4) # Throws error

}
