% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff11_C}
\alias{ff11_C}
\title{A stripped down version of ff1_C. No checking, assumes unit weighted states and integer 1:(node state space dimension) node state names.}
\usage{
ff11_C(ns, dots)
}
\arguments{
\item{ns}{State label}

\item{nss_vec}{Vector for state names. Use in place of dots arguement.}
}
\value{
The feature function, which is a arma unsigned vector.
}
\description{
Built-in feature function for convenience and testing.
}
\details{
A stripped down version of ff1_C. Built for speed....I hope.
}
