% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff12_C}
\alias{ff12_C}
\title{Very stripped down version of ff1_C. No checking, assumes unit weighted states and integer 1:(node state space dimension) node state names.}
\usage{
ff12_C(ns, nss_vec)
}
\arguments{
\item{ns}{State label}

\item{nss_vec}{Vector for state names. Use in place of dots arguement.}
}
\value{
The feature function, which is a arma unsigned vector.
}
\description{
Built-in feature function for convenience and testing.
}
\details{
Very stripped down version of ff1_C. Needs a helper R wrapper (ff12_RC()) because of the whole dots thing.
Built for speed....I hope.
}
