% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff13_C}
\alias{ff13_C}
\title{Very ver stripped down version of ff1_C. No checking, assumes unit weighted states and integer node state names starting at 1.}
\usage{
ff13_C(ns, nss_dim)
}
\arguments{
\item{ns}{State label (integer between 1 and nss_dim)}

\item{nss_dim}{Node state space dimension.}
}
\value{
The feature function, which is an Rcpp integer vector.
}
\description{
Built-in feature function for convenience and testing.
}
\details{
Very very stripped down version of ff1_C. Built for speed. Assumes state space is
integer valued, 1:nss_dim. Probably should use this with an R-side wrapper to sort out other
arguments and arbitrary state names. CAUTION! No checks performed. This function can easily
segfault.
}
