% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff1_times_C}
\alias{ff1_times_C}
\title{Feature function in Rcpp example with timers via RcppClock}
\usage{
ff1_times_C(ns, dots = NULL)
}
\arguments{
\item{ns}{state label of node (\emph{i.e.} node state)}

\item{dots}{Optional arguments: \cr
\describe{
  \item{\code{nss.vec}}{node state space (a vector)}
  \item{\code{nss.dim}}{node state space dimension}
  \item{\code{w.vec}}{node state space weight vector}
}}
}
\value{
The feature function, which is a arma vector.
}
\description{
Built-in feature function for convenience and testing. Works pretty much like ff1 on the R side.
}
\details{
Example function with timers via RcppClock for timing execution on blocks of code.
This is also part of experiments to explore why ff1_C runs 2X-4X as slow as ff0.
}
\examples{
XXXX

}
