% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff2_C}
\alias{ff2_C}
\title{Very stripped down version of ff0_C. No checking and assumes unit weighted states.}
\usage{
ff2_C(st, st_vec)
}
\arguments{
\item{st}{State label}

\item{st_vec}{Vector for state names. Needs to be at least two (numerical) components.}
}
\value{
The feature function, which is a arma unsigned vector.
}
\description{
Built-in feature function for convenience and testing.
}
\details{
Very stripped down version of ff0_C. Runs fastest of the three: ff0_C, ff1_C and ff2_C.
}
