% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_utils.R
\name{get_script}
\alias{get_script}
\title{Function to download script from the Notes}
\usage{
get_script(script_name, url_head = NULL, download_location = NULL)
}
\arguments{
\item{script_name}{Name of the script in the Notes}

\item{url_head}{Location of the script (optional). If NULL, default location is used.}

\item{download_location}{Where to download the script. If NULL, default location is used.}
}
\value{
The feature function, which is a vector.
}
\description{
The function will download script from the Notes
}
\details{
The function will download script from the Notes. Default url_head is currently: XXXX.
Default download_location is currently: XXXX.
}
\examples{
XX

}
