% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{inQ}
\alias{inQ}
\title{Check if element is in a vector}
\usage{
inQ(element, vector)
}
\arguments{
\item{element}{Something to compare to. Can be any single element RObject}

\item{vector}{A vector of things to compare element to. Can be any types.}
}
\value{
A bool. 1 = element in vector, 0 otherwise
}
\description{
Check if element is in a vector, independent of data type.
}
\details{
I'm tired of writing this out longhand in the codes so writing a wrapper.
}
\examples{
XXXX

}
