% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{insert.params.and.pots}
\alias{insert.params.and.pots}
\title{Input parameter vector and potentials.}
\usage{
insert.params.and.pots(crf, params.samp)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
