% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{log_list}
\alias{log_list}
\title{Take the log of stuff (vectors, matrices, etc) in a list}
\usage{
log_list(a.list.for.logging, neglogQ = T)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
XX
}
\details{
Handy for going from node/edge potentials in list form to node/edge energies (log potentials)
}
