% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp}
\alias{logsumexp}
\title{logsumexp trick}
\usage{
logsumexp(logv)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Log sum exp trick. Code ported from Brendon Brewer's DNest code
}
\details{
The function will XXXX
}
