% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp2}
\alias{logsumexp2}
\title{Log sum exp trick from above. A less readable but more compact version:}
\usage{
logsumexp2(logv)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
