% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{make.empty.field}
\alias{make.empty.field}
\title{Instantiate an empty field}
\usage{
make.empty.field(
  graph.eq = NULL,
  adj.mat = NULL,
  num.states = 2,
  state.names = NULL,
  parameterization.typ = "standard",
  node.par = NULL,
  edge.par = NULL,
  plotQ = FALSE
)
}
\arguments{
\item{graph.eq}{XX}

\item{adj.mat}{XX}

\item{num.states}{Scalar or vector specifying the number of states per node.
If a scalar, all nodes assumed to have that number of states.
Minimum magnitude is 2 states per node (Ising-like).}

\item{parameterization.typ}{Type of parameterization for the potentials. Can be general, standard, flexible.}

\item{node.par}{XX}

\item{edge.par}{XX}

\item{plotQ}{XX}
}
\value{
A crf object.
}
\description{
Instantiate an empty field. Should work for Ising and Potts-like models.
}
\details{
Instantiates an empty field. Should work for Ising and Potts-type models under all
parameterizations. Potentials are not filled in however. Use insert.params.and.pots() plunck in
generated random parameters/potentials. Or you can add in your own later. NOTE: this assumes only
one and two-body potentials (n.nf=1, n.ef=1, cf. make.features()).
}
\examples{
XXXX

}
