% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{make.gRbase.potentials}
\alias{make.gRbase.potentials}
\title{Decorate initalized potentials in crf object to be compatible with gRbase}
\usage{
make.gRbase.potentials(crf)
}
\arguments{
\item{crf}{a crf object}
}
\value{
A list with the decorated potentials and their corresponding energies.
}
\description{
XX
}
\details{
Decorates the node and edge potentials with naming conventions useful
with or required by gRbase functionality.
}
\examples{
XXXX

}
