% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_util.R
\name{model.latex}
\alias{model.latex}
\title{Spit out LaTeX for parameter and potential matrices.}
\usage{
model.latex(crf, type = "parameter", symbols = "theta", indices = "vector")
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
The function will XXXX
}
