% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{n2i}
\alias{n2i}
\title{Name(s) to index(ies)}
\usage{
n2i(name.vec, ordered.names)
}
\arguments{
\item{name.vec}{One or more name elements. Usually will be state character names.}

\item{ordered.names}{All possible unique names in a specified order. Usually will be the names of states a node can assume.}
}
\value{
The index of the names in name.vec
}
\description{
Convenience function to convert a character name to an index
}
\details{
Function intended to convert a node state name to a state index for use
with the more efficient energy functions config.energy.e, config_energy_eC, energye
and energyeC. Function should also work to convert node name to node index if
needed. Note however that conversion is in the crf object element node.name.tab.
}
