% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{order.configs}
\alias{order.configs}
\title{Order configurations given node state indices}
\usage{
order.configs(configs.mat, crf, order.nodesQ = F)
}
\arguments{
\item{configs.mat}{A matrix of configurations}

\item{crf}{A crf object}

\item{order.nodesQ}{Whether or not to order the columns (i.e. nodes) in the same order as is found in the crf object.}
}
\value{
A list containing canonically ordered configuration matrix in terms of node state
indices and the re-order indices themselves.
}
\description{
Give a canonical order to a matrix of configurations via the node order (columns) using order()
}
\details{
Orders the configurations column by column in node state order using the R base
function order. Given the states of each node are assigned indices in the crf object, the
function treats each configuration an integer vector and orders them node-wise
(i.e. column-wise). This is handy when comparing distributions with the same configurations and
nodes. By default it's assumed the configurations are sent in already in the column order you
want (order.nodesQ=F). With order.nodesQ=T the columns will be first ordered according to whats
in crf$node.name.tab and then the configurations (rows) will be ordered. If order.nodesQ=T the
column rearrangment indices will also be returned.
}
