% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{phi.features.e.C}
\alias{phi.features.e.C}
\title{Wrapper for Rcpp version of phi features}
\usage{
phi.features.e.C(config, crf)
}
\arguments{
\item{config}{a configuration}

\item{crf}{a crf object}
}
\value{
a phi vector
}
\description{
Wrapper for Rcpp version of phi features
}
\details{
Wrapper for Rcpp version of phi features. This version assumes state of the configuration are
node state indices.
}
\examples{
XXXX

}
