% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{phi.features.e}
\alias{phi.features.e}
\title{Compute features phi faster}
\usage{
phi.features.e(config, crf)
}
\arguments{
\item{config}{a configuration}

\item{crf}{a crf object}
}
\value{
a phi vector
}
\description{
Utility function to compute unconditional features phi_i(config)
}
\details{
Assumes feature components phi_i(X), are 0,1 valued and parameters are contiguous and
consecutively numbered. More importantly, this version assume state of the configuration are
node state indices. That's how we avoid the slower matrix multiplies and replace them with
lookups.
}
\examples{
XXXX

}
