% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{phi.features}
\alias{phi.features}
\title{Compute features phi}
\usage{
phi.features(config, edges.mat, node.par, edge.par, ff)
}
\arguments{
\item{config}{configuration}

\item{edges.mat}{an edge index map for a graph}

\item{node.par}{node parameter index matrix for a graph}

\item{edge.par}{list of edge parameter index matries for a graph}

\item{ff}{feature function}
}
\value{
a phi vector
}
\description{
Utility function to compute unconditional features phi_i(config)
}
\details{
Assumes feature components phi_i(X), are 0,1 valued and parameters are numbered. Any
names for the states of the configuration are allowed so long as they are consistent with the
feature functions. This version is slower however.
}
\examples{
XXXX

}
