% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot.configuration.sample}
\alias{plot.configuration.sample}
\title{Plot configuration sample}
\usage{
\method{plot}{configuration.sample}(samples, crf = NULL, num.top.configs = 3)
}
\arguments{
\item{samples}{a crf object}
}
\value{
Nothing.
}
\description{
Plot the sample of configurations
}
\details{
The function will plot the counts of each unique configuration of states in a sample of configurations.
}
\examples{
XXXX

}
