% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plot.marginal.sample2}
\alias{plot.marginal.sample2}
\title{Plot marginal sample}
\usage{
\method{plot}{marginal.sample2}(samples, crf)
}
\arguments{
\item{samples}{The samples}
}
\value{
Nothing.
}
\description{
Plot the sample of states at each node
}
\details{
The function will plot the counts of each state at each node.
}
\examples{
XXXX

}
