% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_prototypef.R
\name{prototype_func}
\alias{prototype_func}
\title{Prototype function}
\usage{
prototype_func(avar)
}
\arguments{
\item{avar}{A variable}
}
\value{
Some stuff.
}
\description{
The function is handy to cut and paste for the roxygen header.
}
\details{
The function is an example to cut and past quick.
}
