% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{row.match}
\alias{row.match}
\title{row.match in a matrix or table}
\usage{
row.match(x, table, nomatch = NA)
}
\arguments{
\item{XX}{}
}
\value{
The function will XX
}
\description{
Code from prodlim library to match a row in a matrix
}
\details{
The function will XXXX
}
