% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{row_match}
\alias{row_match}
\title{row.match in C}
\usage{
row_match(x, table)
}
\arguments{
\item{x}{a row vector}

\item{table}{a matrix to be searched for rows matching x}
}
\value{
Indices of matching rows
}
\description{
Looks for matching rows in a matrix.
}
\details{
Much faster than in R and we don't want to pass in R functions anyway.
}
\examples{
XXXX

}
