% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRbase_utils.R
\name{tabNew.crf}
\alias{tabNew.crf}
\title{Slightly augmented version of gRbase tabNew to be used directly with a crf object}
\usage{
tabNew.crf(node.name.vec, vals, crf, ...)
}
\arguments{
\item{node.name.vec}{node names for the gRbase decorated table}

\item{vals}{values to go into gRbase decorated table}

\item{crf}{a crf object}

\item{...}{extra arguments to tabNew}
}
\value{
XX
}
\description{
Slightly augmented version of gRbase tabNew to be used in making new tables directly with a crf object
}
\details{
A slightly augmented version of gRbase tabNew which will pull the node state names out
of the crf object and set everything up fairly automatically to make a new, properly decorated
table that is compatible with gRbase functions. The input values can be in the form of a
data frame (e.g. a bunch of configurations), a table which is a contingency table or just a
vector of values.
}
\examples{
XX

}
