% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_CRFutil2}
\alias{update_CRFutil2}
\title{Easily update the CRFutil2 library by installing the current version from the github site}
\usage{
update_CRFutil2()
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Easily update the CRFutil2 library by installing the current version from the github site
}
\details{
Easily update the CRFutil2 library by installing the current version from the github site
}
\examples{
XX

}
