% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dress.sample}
\alias{dress.sample}
\title{Dress a sample with node and state names in crf object}
\usage{
dress.sample(crf, samples)
}
\arguments{
\item{crf}{A crf object which also contains node and state names (i.e. out augmented crf object)}

\item{samples}{A set of configurations with state indices as names. Typically these will be a samples generated by one of the CRF sampling functions.}
}
\value{
The sample dressed with node and state names contained in the crf object.
}
\description{
Dress a sample or matrix of configurations with node and state names in crf object
}
\details{
NOTE: The function takes a configuration matrix of configurations with state
indices as names. The function takes this "index" configuration matrix and dresses it up with
node names as the columns and state names for the configuration outcome elements (the name info
is contained in the crf object). The sampling functions from CRF (e.g. sample.junction()) just
spit out an undressed matrix of samples with sample state numbers (e.g. 1, 2, ....) for the
configuration elements. This is needed sometimes and often more convenient, but for display
purposes the user may want more labels on the sample space. Hence this (cosmetic) function. The
inverse of this function is configs.n2i().
}
\examples{
XXXX

}
