% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy_util.R
\name{Eone}
\alias{Eone}
\title{One-body (node) energy function}
\usage{
Eone(yA, tA, ff)
}
\arguments{
\item{yA}{A node state}

\item{tA}{Tau vector}

\item{ff}{Feature function}
}
\value{
One-body (node) energy.
}
\description{
One-body (node) energy function
}
\details{
One-body (node) energy function
}
\examples{
XXXXX

}
