% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ff1_C}
\alias{ff1_C}
\title{Stripped down version of ff0_C. No checking. Not idiot proof.}
\usage{
ff1_C(st, ss_dim, w_vec, st_vec)
}
\arguments{
\item{st}{State label}

\item{ss_dim}{State space dimension. Minimum is 2 (Ising), but function doesn't check.}

\item{w_vec}{Weight vector. Needs to be of length ss_dim.}

\item{st_vec}{Vector for state names. Needs to be of length ss_dim.}
}
\value{
The feature function, which is a arma vector.
}
\description{
Built-in feature function for convenience and testing.
}
\details{
Stripped down version of ff0_C. No checking. Not idiot proof. Meant to be faster. Runs about as fast as ff0 on the R side.
}
