% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{insert.params.and.pots}
\alias{insert.params.and.pots}
\title{Input a parameter vector into parameter and potential matrices and vectors of crf object}
\usage{
insert.params.and.pots(crf, params.samp)
}
\arguments{
\item{crf}{a crf object}

\item{param.samp}{a sample of parameter space or parameter vector}
}
\value{
Nothing.
}
\description{
XX
}
\details{
Input a parameter vector into parameter and potential matrices of crf object
}
\examples{
XXXX

}
