% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{logsumexp2}
\alias{logsumexp2}
\title{logsumexp trick}
\usage{
logsumexp2(logv)
}
\arguments{
\item{XX}{XX}
}
\value{
The function will XX
}
\description{
Log sum exp trick also. A less readable but more compact version:
}
\details{
The function will XXXX
}
