% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R
\name{make.gRbase.potentials}
\alias{make.gRbase.potentials}
\title{Decorate initalized mrf-object to make potentials compatible with gRbase}
\usage{
make.gRbase.potentials(crf)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
Decorate initalized mrf-object to make potentials compatible with gRbase
}
\details{
The function will XXXX
}
