% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crf_util.R, R/gRbase_utils.R
\name{make.gRbase.potentials}
\alias{make.gRbase.potentials}
\title{Decorate initalized potentials in crf object to be compatible with gRbase}
\usage{
make.gRbase.potentials(crf)

make.gRbase.potentials(crf)
}
\arguments{
\item{crf}{a crf object}

\item{XX}{The XX}
}
\value{
A list with the decorated potentials and their corresponding energies.

The function will XX
}
\description{
XX

Decorate initalized mrf-object to make potentials compatible with gRbase
}
\details{
Decorates the node and edge potentials with naming conventions useful
with or required by gRbase functionality.

The function will XXXX
}
\examples{
XXXX

}
