% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gRbase_utils.R
\name{make.gRbase.potentials}
\alias{make.gRbase.potentials}
\title{Decorate initialized crf-object to make potentials compatible with gRbase}
\usage{
make.gRbase.potentials(crf, includeQ = T)
}
\arguments{
\item{crf}{a crf object with the potentials containing something}

\item{includeQ}{whether or not to write gRbase formatted potentials and energies to the crf object}
}
\value{
See details
}
\description{
Decorate initialized crf-object to make potentials and energies compatible with gRbase
}
\details{
Decorate initialized crf-object potentials to make compatible with gRbase functions. Also
properly decorated energies (negative log potentials) are generated. If the user desires to include
the decorated potentials/energies in the crf object, set includeQ=T (the default). They will be
included in a section labeled gR. If includeQ=F the potentials and energies are simply returned
as a list.
}
\examples{
XX

}
