% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{row.match}
\alias{row.match}
\title{Code from prodlim library to match a row in a matrix}
\usage{
row.match(x, table, nomatch = NA)
}
\arguments{
\item{XX}{The XX}
}
\value{
The function will XX
}
\description{
XXXX
}
\details{
The function will XXXX
}
